+++ src/programs/package.ts
@@ -1,3 +1,4 @@
+import camelcase from "camelcase";
 import * as path from "node:path";
 import ts from "typescript";
 import * as yargs from "yargs";
@@ -80,7 +81,9 @@ async function main(options: MainOptions) {
     const nodes = context.getNodes();
 
     const namer = new Namer(options.uniqueNameSeed);
-    for (const [nodeId, typeName] of context.getTypeNames()) {
+    for (const nodeId of Object.keys(nodes)) {
+        const nodeUrl = new URL(nodeId);
+        const typeName = getNodeTypeName(nodeUrl);
         namer.registerName(nodeId, typeName);
     }
 
@@ -92,4 +95,33 @@ async function main(options: MainOptions) {
         name: packageName,
         version: packageVersion,
     });
+}
+
+function getNodeTypeName(nodeUrl: URL): string {
+    const reReplace = /[^A-Za-z0-9-_.,]/gu;
+
+    const pointer = nodeUrl.hash.startsWith("#") ? nodeUrl.hash.substring(1) : "";
+
+    const pathParts = nodeUrl.pathname
+        .split("/")
+        .map(decodeURI)
+        .map((value) => value.replace(reReplace, ""))
+        .filter((value) => value !== "");
+    const pointerParts = pointer
+        .split("/")
+        .map(decodeURI)
+        .map((value) => value.replace(reReplace, ""));
+
+    const nameParts = [
+        pathParts[pathParts.length - 1] ?? "Schema",
+        pointerParts[pointerParts.length - 3],
+        pointerParts[pointerParts.length - 2],
+        pointerParts[pointerParts.length - 1],
+    ]
+        .filter((value) => value != null)
+        .filter((value) => value != "");
+
+    const name = camelcase(nameParts, { pascalCase: true });
+
+    return name;
 }
